requirements.txt
python-telegram-bot==21.4
requests==2.31.0
python-dotenv==1.0.1
dotenv==0.9.9
transliterate==1.10.2
Unidecode==1.3.8